Build universal iOS framework
Most importantly, this exposes the native library as a "framework" in iOS so that a consuming iOS application will not be rejected by the App Store.
In addition to this though, we combine both x64 and arm64 binaries into one *universal* binary, which should help make things "just work" whether in an emulator or a real device.